sweetalert 2 and alert when delete client

+++ src/app/pages/client/client.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CrudService } from '../../services/crud/crud.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-client',
@@ -36,11 +37,29 @@ export class ClientComponent implements OnInit {
   }
 
   deleteClient(id: number) {
-    this.crudService.delete(this.model, id).
-      subscribe((response) => {
-        //return true or false
-        this.getData();
-      });
+    //TODO if response true, then show swal.fire
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.crudService.delete(this.model, id).
+          subscribe((response) => {
+            //return true or false
+            this.getData();
+          });
+        Swal.fire(
+          'Deleted!',
+          'Client has been deleted.',
+          'success'
+        );
+      }
+    });
   }
 
 }
